Refactor (event-tool): Try -> Result, split services up, & use correct coroutine scope
