+++ packages/wds-event-tool/build.gradle.kts
@@ -11,6 +11,7 @@ import org.jooq.util.jaxb.Strategy
 import org.jooq.util.jaxb.Target
 
 val kotlinVersion = "1.3.61"
+val coroutinesVersion = "1.3.3"
 val http4kVersion = "3.235.0"
 val log4jVersion = "2.13.0"
 val jacksonVersion = "2.10.1"
@@ -18,8 +19,7 @@ val firebaseVersion = "6.6.0"
 val flywayCoreVersion = "5.2.4"
 val postgresqlDriverVersion = "42.2.5"
 val jooqVersion = "3.10.1"
-val arrowVersion = "0.8.2"
-// val newArrowVersion = "0.10.4"
+val arrowVersion = "0.10.4"
 val mockkVersion = "1.9.2"
 val KGraphQLVersion = "0.6.4"
 
@@ -71,7 +71,7 @@ dependencies {
     /* kotlin */
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlinVersion")
     implementation("org.jetbrains.kotlin:kotlin-reflect:$kotlinVersion")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.1.0")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion")
     /* http4k */
     implementation("org.http4k:http4k-core:$http4kVersion")
     implementation("org.http4k:http4k-contract:$http4kVersion")
@@ -105,12 +105,9 @@ dependencies {
     /* Result */
     implementation("com.michael-bull.kotlin-result:kotlin-result:1.1.6")
     /* Arrow new */
-    // implementation("io.arrow-kt:arrow-fx:$newArrowVersion")
-    // implementation("io.arrow-kt:arrow-mtl:$newArrowVersion")
-    // implementation("io.arrow-kt:arrow-syntax:$newArrowVersion")
-    /* Arrow old */
-    implementation("io.arrow-kt:arrow-core:$arrowVersion")
-    implementation("io.arrow-kt:arrow-data:$arrowVersion")
+    implementation("io.arrow-kt:arrow-fx:$arrowVersion")
+    implementation("io.arrow-kt:arrow-mtl:$arrowVersion")
+    implementation("io.arrow-kt:arrow-syntax:$arrowVersion")
 
     /* Validation */
     compile("com.markodevcic.kvalidation:KValidation:1.0.0")
