+++ packages/wds-event-tool/src/main/kotlin/org/webdevandsausages/events/controller/AdminGetEventInfo.kt
@@ -12,7 +12,7 @@ import org.http4k.core.Status
 import org.webdevandsausages.events.Router
 import org.webdevandsausages.events.dto.EventDto
 import org.webdevandsausages.events.domain.toResponse
-import org.webdevandsausages.events.service.GetEventByIdService
+import org.webdevandsausages.events.service.event.GetEventByIdService
 import org.webdevandsausages.events.utils.WDSJackson.auto
 
 object AdminGetEventInfo {
