+++ packages/wds-event-tool/src/main/kotlin/org/webdevandsausages/events/controller/GetCurrentEvent.kt
@@ -11,7 +11,7 @@ import org.http4k.core.Status
 import org.webdevandsausages.events.ApiRouteWithGraphqlConfig
 import org.webdevandsausages.events.dto.EventOutDto
 import org.webdevandsausages.events.domain.toResponse
-import org.webdevandsausages.events.service.GetCurrentEventService
+import org.webdevandsausages.events.service.event.GetCurrentEventService
 import org.webdevandsausages.events.utils.WDSJackson.auto
 
 object GetCurrentEvent: ApiRouteWithGraphqlConfig {
