+++ packages/wds-event-tool/src/main/kotlin/org/webdevandsausages/events/controller/GetUser.kt
@@ -13,7 +13,7 @@ import org.webdevandsausages.events.ApiRouteWithGraphqlConfig
 import org.webdevandsausages.events.Router
 import org.webdevandsausages.events.dto.RegistrationOutDto
 import org.webdevandsausages.events.domain.toResponse
-import org.webdevandsausages.events.service.GetRegistrationService
+import org.webdevandsausages.events.service.registration.GetRegistrationService
 
 object GetUser : ApiRouteWithGraphqlConfig {
     private var regService: GetRegistrationService? = null
