+++ packages/wds-event-tool/src/main/kotlin/org/webdevandsausages/events/controller/PostRegistration.kt
@@ -1,5 +1,6 @@
 package org.webdevandsausages.events.controller
 
+import CreateRegistrationService
 import com.apurebase.kgraphql.schema.dsl.SchemaBuilder
 import org.http4k.contract.ContractRoute
 import org.http4k.contract.div
@@ -19,7 +20,6 @@ import org.webdevandsausages.events.dto.ErrorCode
 import org.webdevandsausages.events.dto.ErrorOutDto
 import org.webdevandsausages.events.dto.RegistrationInDto
 import org.webdevandsausages.events.dto.RegistrationOutDto
-import org.webdevandsausages.events.service.CreateRegistrationService
 import org.webdevandsausages.events.utils.Read
 import org.webdevandsausages.events.utils.WDSJackson.auto
 import org.webdevandsausages.events.utils.parse
