+++ packages/wds-event-tool/src/main/kotlin/org/webdevandsausages/events/service/FirebaseService.kt
@@ -1,20 +1,23 @@
 package org.webdevandsausages.events.service
 
-import arrow.core.Try
+import com.github.michaelbull.result.fold
+import com.github.michaelbull.result.runCatching
 import com.google.auth.oauth2.GoogleCredentials
 import com.google.cloud.firestore.SetOptions
 import com.google.firebase.FirebaseApp.initializeApp
 import com.google.firebase.FirebaseOptions
 import com.google.firebase.cloud.FirestoreClient
-import org.slf4j.LoggerFactory
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import org.webdevandsausages.events.dto.ParticipantDto
 import org.webdevandsausages.events.dto.participantDetails
+import org.webdevandsausages.events.utils.createLogger
 import java.io.InputStream
 
 fun String.asResourceStream(): InputStream? = Thread.currentThread().contextClassLoader.getResourceAsStream(this)
 
-class FirebaseService {
-    val logger = LoggerFactory.getLogger("firebase")
+class FirebaseService : CoroutineScope by CoroutineScope(Dispatchers.Default) {
+    val logger = createLogger()
 
     init {
         val serviceAccount = "firebaseServiceAccountKey.json".asResourceStream()
@@ -33,12 +36,13 @@ class FirebaseService {
         val details = participant.participantDetails
         val email = details["email"]
         if (email != null) {
-            Try { participantsRef?.document(email)?.set(details, SetOptions.merge()) }.fold({
-                logger?.error("Firebase failed to save $email to mailing list: ${it.message}")
-            },
-            {
-                logger?.info("Firebase saved $email to mailing list.")
-            })
+            runCatching { participantsRef?.document(email)?.set(details, SetOptions.merge()) }.fold(
+                {
+                    logger.info("Firebase saved $email to mailing list.")
+                },
+                {
+                    logger.error("Firebase failed to save $email to mailing list: ${it.message}")
+                })
         }
     }
 }
