+++ packages/wds-event-tool/src/main/kotlin/org/webdevandsausages/events/service/event/GetEventsService.kt
@@ -1,0 +1,10 @@
+package org.webdevandsausages.events.service.event
+
+import org.webdevandsausages.events.dao.EventCRUD
+import org.webdevandsausages.events.dto.EventDto
+
+class GetEventsService(val eventRepository: EventCRUD) {
+    operator fun invoke(status: String?): List<EventDto> {
+        return eventRepository.findAllWithParticipants(status)
+    }
+}
